Much progress on track fingerprinting!
* Adding libofa and its dependencies (some of libfftw, libcurl and libexpat) to the repo. I'm using libofa to generate a fingerprint based on segments of tracks. The overall process I envision is:
       1.) Decompress a track to PCM (my test tracks are all mp3) using the Foobar SDK.
       2.) Extract the samples and other data libofa requires.
       3.) Send the existing metadata and samples to the MusicDNS server which attempts to fuzzy match it with existing samples. This (ideally) returns a PUID and metadata.
       4.) Query MusicBrainZ for all the tracks with that PUID.

I have step 1-3 working quite nicely. Unfortunately, it seems to be annoyingly difficult to use libmusicbrainz to do Step 4 - I might be able to hack together my own XML query to do the job because MusicBrainZ does publicly offer this sort of query on their site (but for some reason not in their API). Otherwise, I can use the metadata returned by MusicDNS to narrow down the candidate matches.

The downside of all of this is that it makes deployment increasingly more difficult... now I need to deploy DLLs for libofa and its 3 dependencies as DLLs with my plugin. But I'll attempt to statically link against libfftw3 instead (since the DLL is a meg and a half). I already have neon as a dependency so I might be able to replace curl with it. libexpat is just there for XML parsing and I have a bazillion options there. Watch this space.

* Upgrading libmusicbrainz from 3.0.2 to 3.0.3. As I understand it, this adds some Release Group related functionality which will be needed.

* Upgrading plan.txt now that my ideas for the matching heuristics are crystallising.